chore: move FeeMultiplierProvider to traits
