+++ state-chain/runtime/src/configs.rs
@@ -37,7 +37,8 @@ pub use cf_traits::{
 	AccountInfo, Chainflip, EpochInfo, PoolApi, QualifyNode, SessionKeysRegistered, SwappingApi,
 };
 use cf_traits::{
-	ChainflipWithTargetChain, DummyEgressSuccessWitnesser, DummyIngressSource, NoLimit,
+	ChainflipWithTargetChain, DefaultFeeMultiplier, DummyEgressSuccessWitnesser,
+	DummyIngressSource, NoLimit,
 };
 
 pub use frame_support::{
@@ -1014,37 +1015,37 @@ impl pallet_cf_asset_balances::Config for Runtime {
 
 impl pallet_cf_chain_tracking::Config<Instance1> for Runtime {
 	type TargetChain = Ethereum;
-	type FeeMultiplierProvider = pallet_cf_chain_tracking::DefaultFeeMultiplier;
+	type FeeMultiplierProvider = DefaultFeeMultiplier;
 	type WeightInfo = pallet_cf_chain_tracking::weights::PalletWeight<Runtime>;
 }
 
 impl pallet_cf_chain_tracking::Config<Instance2> for Runtime {
 	type TargetChain = Polkadot;
-	type FeeMultiplierProvider = pallet_cf_chain_tracking::DefaultFeeMultiplier;
+	type FeeMultiplierProvider = DefaultFeeMultiplier;
 	type WeightInfo = pallet_cf_chain_tracking::weights::PalletWeight<Runtime>;
 }
 
 impl pallet_cf_chain_tracking::Config<Instance3> for Runtime {
 	type TargetChain = Bitcoin;
-	type FeeMultiplierProvider = pallet_cf_chain_tracking::DefaultFeeMultiplier;
+	type FeeMultiplierProvider = DefaultFeeMultiplier;
 	type WeightInfo = pallet_cf_chain_tracking::weights::PalletWeight<Runtime>;
 }
 
 impl pallet_cf_chain_tracking::Config<Instance4> for Runtime {
 	type TargetChain = Arbitrum;
-	type FeeMultiplierProvider = pallet_cf_chain_tracking::DefaultFeeMultiplier;
+	type FeeMultiplierProvider = DefaultFeeMultiplier;
 	type WeightInfo = pallet_cf_chain_tracking::weights::PalletWeight<Runtime>;
 }
 
 impl pallet_cf_chain_tracking::Config<Instance5> for Runtime {
 	type TargetChain = Solana;
-	type FeeMultiplierProvider = pallet_cf_chain_tracking::DefaultFeeMultiplier;
+	type FeeMultiplierProvider = DefaultFeeMultiplier;
 	type WeightInfo = pallet_cf_chain_tracking::weights::PalletWeight<Runtime>;
 }
 
 impl pallet_cf_chain_tracking::Config<Instance6> for Runtime {
 	type TargetChain = Assethub;
-	type FeeMultiplierProvider = pallet_cf_chain_tracking::DefaultFeeMultiplier;
+	type FeeMultiplierProvider = DefaultFeeMultiplier;
 	type WeightInfo = pallet_cf_chain_tracking::weights::PalletWeight<Runtime>;
 }
